upd(docker-compose): Upgrade PostgreSQL image to 15 and update volumes, commands, and healthchecks.
This commit updates the PostgreSQL image to version 15, changes the volume mounting for the database and core services, adds a command to log all statements in PostgreSQL, and updates healthcheck settings for the database service. Additionally, it removes the unused volumes section at the end of the file. These changes aim to improve the overall performance and logging of the application.